requirements.txt
click==8.1.7
pycobol==0.1.0
openai==1.97.1
anthropic==0.59.0
requests==2.31.0
jinja2==3.1.2
pyyaml==6.0.1
colorama==0.4.6
tqdm==4.66.1
networkx==3.2.1
matplotlib==3.10.3
plotly==5.17.0
pandas==2.3.1
numpy==2.2.6 